+++ src/api/domain/decks/deck.controller.spec.ts
@@ -170,4 +170,158 @@ describe('[API] Deck controller', () => {
     expect(res.statusCode).toEqual(201);
     expect(typeof res.body.deckId).toEqual('string');
   });
+
+  test('[PUT] /decks/update-name - should return error if data is invalid', async () => {
+    process.env.JWT_TOKEN = 'secret';
+
+    emailChecker.isUnique.mockResolvedValue(true);
+
+    const email = faker.internet.email();
+    const password = faker.internet.password();
+    const username = faker.name.findName();
+
+    const user = await UserRegistration.registerNew(
+      {
+        email,
+        password,
+        username,
+      },
+      emailChecker,
+    );
+
+    await userRegistrationRepository.insert(user);
+
+    const token = jwt.sign(
+      {
+        userId: user.getId().getValue(),
+      },
+      process.env.JWT_TOKEN,
+    );
+
+    const res = await request(app)
+      .put('/decks/update-name')
+      .set('Accept', 'application/json')
+      .set('Authorization', `Bearer ${token}`);
+
+    expect(res.statusCode).toEqual(422);
+    expect(res.body.details.map((detail) => detail.key)).toEqual(['newName', 'deckId']);
+  });
+
+  test('[PUT] /decks/update-name - should return error if user is not authorized', async () => {
+    const res = await request(app).put('/decks/update-name').set('Accept', 'application/json');
+
+    expect(res.statusCode).toEqual(401);
+    expect(res.body.error).toEqual('Unauthorized.');
+  });
+
+  test('[PUT] /decks/update-name - should return error if deck name is not unique', async () => {
+    process.env.JWT_TOKEN = 'secret';
+
+    emailChecker.isUnique.mockResolvedValue(true);
+
+    const email = faker.internet.email();
+    const password = faker.internet.password();
+    const username = faker.name.findName();
+
+    const user = await UserRegistration.registerNew(
+      {
+        email,
+        password,
+        username,
+      },
+      emailChecker,
+    );
+
+    await userRegistrationRepository.insert(user);
+
+    const token = jwt.sign(
+      {
+        userId: user.getId().getValue(),
+      },
+      process.env.JWT_TOKEN,
+    );
+
+    const name = faker.name.findName();
+
+    const deck = Deck.instanceExisting(
+      {
+        name,
+        cardIDs: [],
+        createdAt: new Date(),
+        description: '#description',
+        ownerId: user.getId(),
+        tags: ['#tag-1'],
+      },
+      new UniqueEntityID(),
+    );
+
+    await deckRepository.insert(deck);
+
+    const res = await request(app)
+      .put('/decks/update-name')
+      .set('Accept', 'application/json')
+      .set('Authorization', `Bearer ${token}`)
+      .send({
+        newName: name,
+        deckId: deck.getId().getValue(),
+      });
+
+    expect(res.statusCode).toEqual(400);
+    expect(res.body.error).toEqual(`You've already created deck with name: "${name}".`);
+  });
+
+  test('[PUT] /decks/update-name - should update deck name', async () => {
+    process.env.JWT_TOKEN = 'secret';
+
+    emailChecker.isUnique.mockResolvedValue(true);
+
+    const email = faker.internet.email();
+    const password = faker.internet.password();
+    const username = faker.name.findName();
+
+    const user = await UserRegistration.registerNew(
+      {
+        email,
+        password,
+        username,
+      },
+      emailChecker,
+    );
+
+    await userRegistrationRepository.insert(user);
+
+    const token = jwt.sign(
+      {
+        userId: user.getId().getValue(),
+      },
+      process.env.JWT_TOKEN,
+    );
+
+    const name = faker.name.findName();
+
+    const deck = Deck.instanceExisting(
+      {
+        name,
+        cardIDs: [],
+        createdAt: new Date(),
+        description: '#description',
+        ownerId: user.getId(),
+        tags: ['#tag-1'],
+      },
+      new UniqueEntityID(),
+    );
+
+    await deckRepository.insert(deck);
+
+    const res = await request(app)
+      .put('/decks/update-name')
+      .set('Accept', 'application/json')
+      .set('Authorization', `Bearer ${token}`)
+      .send({
+        newName: faker.name.findName(),
+        deckId: deck.getId().getValue(),
+      });
+
+    expect(res.statusCode).toEqual(204);
+  });
 });