+++ src/core/decks/deck-tag/deck-tag.repository.ts
@@ -2,4 +2,6 @@ import { DeckTag } from './deck-tag.entity';
 
 export interface DeckTagRepository {
   insert(deckTag: DeckTag): Promise<void>;
+
+  findAllByDeck(deckId: string): Promise<DeckTag[]>;
 }