+++ src/core/decks/decks/deck.repository.ts
@@ -2,4 +2,8 @@ import { Deck } from './deck.aggregate-root';
 
 export interface DeckRepository {
   insert(deck: Deck): Promise<void>;
+
+  update(deck: Deck): Promise<void>;
+
+  findById(id: string): Promise<Deck | null>;
 }