+++ src/infrastructure/domain/decks/deck/deck.mapper.ts
@@ -1,4 +1,5 @@
 import { Deck } from '@core/decks/decks/deck.aggregate-root';
+import { UniqueEntityID } from '@core/shared/unique-entity-id';
 
 interface DeckRecord {
   id: string;
@@ -21,5 +22,23 @@ export class DeckMapper {
       created_at: deck.getCreatedAt().toISOString(),
       owner_id: deck.getOwnerId().getValue(),
     };
+  }
+
+  public static toEntity(
+    { id, created_at, owner_id, image_url, ...record }: DeckRecord,
+    cardIDs: UniqueEntityID[] = [],
+    tags: string[] = [],
+  ): Deck {
+    return Deck.instanceExisting(
+      {
+        ...record,
+        cardIDs,
+        createdAt: new Date(created_at),
+        ownerId: new UniqueEntityID(owner_id),
+        tags,
+        imageUrl: image_url,
+      },
+      new UniqueEntityID(id),
+    );
   }
 }